Restructure code in zeta-cli
